requirements.txt
binance==3.1.1
python-dotenv==1.0.0
aiohttp==3.9.3
async-timeout==4.0.3